敵およびプレイヤーのステータスを開く処理をリファクタリングし、専用の関数を作成
