+++ src/js/td-scenes/battle/procedure/open-player-status.ts
@@ -1,0 +1,28 @@
+import { StatusDialog } from "../../../dom-dialogs/status";
+import { BattleSceneProps } from "../props";
+import { switchStatus } from "./switch-status";
+
+/**
+ * プレイヤーのステータスを開く
+ * @param props プロパティ
+ */
+export function openPlayerStatus(props: Readonly<BattleSceneProps>): void {
+  const { stateHistory, playerId } = props;
+  const lastState = stateHistory.at(-1);
+  if (!lastState) {
+    return;
+  }
+
+  const player = lastState.players.find((p) => p.playerId === playerId);
+  if (!player) {
+    return;
+  }
+
+  const dialog = new StatusDialog({
+    ...props,
+    state: player,
+    isEnemy: false,
+    isPilotHidden: false,
+  });
+  switchStatus(props, dialog);
+}
